:hammer: refactor: 타입선언 리팩토링
 - enum 타입이 글로벌로 선언되어 발생하는 문제 해결
