+++ src/api/mall/index.ts
@@ -5,8 +5,10 @@ import request from 'api/core';
 const mall = {
     getMall: (): Promise<AxiosResponse> =>
         request({ method: 'GET', url: '/malls' }),
+
     getMallPartners: (): Promise<AxiosResponse> =>
         request({ method: 'GET', url: '/malls/partners' }),
+
     getSslInfo: (): Promise<AxiosResponse> =>
         request({ method: 'GET', url: '/malls/ssl' }),
 };
