+++ src/api/product/product.ts
@@ -1,119 +1,19 @@
 import { AxiosResponse } from 'axios';
 
 import request, { defaultHeaders } from 'api/core';
-
-export enum CRITERION {
-    RECENT_PRODUCT = 'RECENT_PRODUCT',
-}
-
-export enum DIRECTION {
-    DESC_DELIVERY_FEIGN_CLIENT = 'DESCDeliveryFeignClient',
-    DESC = 'DESC',
-    ASC = 'ASC',
-}
-
-export enum SALE_STATUS {
-    ON_SALE = 'ONSALE',
-    ALL_CONDITIONS = 'ALL_CONDITIONS',
-    READY_ON_SALE = 'READY_ONSALE',
-    RESERVATION_AND_ON_SALE = 'RESERVATION_AND_ONSALE',
-}
-
-export enum DISCOUNTED_COMPARISON {
-    GT = 'GT',
-    LTE = 'LTE',
-    GTE = 'GTE',
-    EQ = 'EQ',
-    BETWEEN = 'BETWEEN',
-}
-
-export enum DELIVERY_CONDITION_TYPE {
-    FREE = 'FREE',
-    CONDITIONAL = 'CONDITIONAL',
-    FIXED_FEE = 'FIXED_FEE',
-}
-
-export enum BY {
-    POPULAR = 'POPULAR',
-    SALE_YMD = 'SALE_YMD',
-    DISCOUNTED_PRICE = 'DISCOUNTED_PRICE',
-    REVIEW = 'REVIEW',
-    SALE_CNT = 'SALE_CNT',
-    RECENT_PRODUCT = 'RECENT_PRODUCT',
-    MD_RECOMMEND = 'MD_RECOMMEND',
-    LIKE_CNT = 'LIKE_CNT',
-}
-
-export enum SHIPPING_AREA_TYPE {
-    PARTNER = 'PARTNER',
-    MALL = 'MALL',
-}
-
-export enum CHANNEL_TYPE {
-    NAVER_EP = 'NAVER_EP',
-    DANAWA = 'DANAWA',
-    ENURI = 'ENURI',
-    WONDER = 'WONDER',
-    COOCHA = 'COOCHA',
-    FACEBOOK = 'FACEBOOK',
-}
-
-interface ProductsParams {
-    deliveryTemplateNo: number;
-    hasOptionValues?: boolean;
-    pageSize?: number;
-    pageNumber?: number;
-    productSort?: {
-        criterion?: CRITERION;
-        direction?: DIRECTION;
-    };
-}
-
-interface RestockParams {
-    optionNos: number[];
-    privacyInfoAgreement: boolean;
-    name: string;
-    phone: string;
-}
-
-interface GroupCodeParams {
-    groupManagementCodes: string[];
-    saleStatus?: SALE_STATUS;
-    isSoldOut?: boolean;
-}
-interface ProductSearchParams {
-    filter?: {
-        discountedPrices?: number;
-        keywords?: string;
-        keywordInResult?: string;
-        discountedComparison?: DISCOUNTED_COMPARISON;
-        deliveryConditionType?: DELIVERY_CONDITION_TYPE;
-        saleStatus?: SALE_STATUS;
-        soldout?: boolean;
-        totalReviewCount?: boolean;
-        familyMalls?: boolean;
-        productManagementCd?: string;
-        excludeMallProductNo?: number;
-        includeMallProductNo?: number;
-    };
-    order?: {
-        by?: BY;
-        direction?: DIRECTION;
-        soldoutPlaceEnd?: boolean;
-    };
-    categoryNos?: number;
-    brandNos?: number;
-    partnerNo?: number;
-    clientKey?: string;
-    pageNumber?: number;
-    pageSize?: number;
-    onlySaleProduct?: boolean;
-    hasMaxCouponAmt?: boolean;
-    hasTotalCount?: boolean;
-    hasOptionValues?: boolean;
-    includeSummaryInfo?: boolean;
-    shippingAreaType?: SHIPPING_AREA_TYPE;
-}
+import {
+    CHANNEL_TYPE,
+    CRITERION,
+    ORDER_DIRECTION,
+    PRODUCT_DIRECTION,
+    PRODUCT_SALE_STATUS,
+} from 'models';
+import {
+    GroupCodeParams,
+    ProductSearchParams,
+    ProductsParams,
+    RestockParams,
+} from 'models/product';
 
 const product = {
     // TODO deliveryTemplateNo을 모름 500 error 발생 추후 테스트 필요
@@ -124,7 +24,7 @@ const product = {
         pageNumber = 1,
         productSort = {
             criterion: CRITERION.RECENT_PRODUCT,
-            direction: DIRECTION.DESC_DELIVERY_FEIGN_CLIENT,
+            direction: PRODUCT_DIRECTION.DESC_DELIVERY_FEIGN_CLIENT,
         },
     }: ProductsParams): Promise<AxiosResponse> =>
         request({
@@ -152,7 +52,7 @@ const product = {
 
     getGroupManagementCodes: ({
         groupManagementCodes,
-        saleStatus = SALE_STATUS.ON_SALE,
+        saleStatus = PRODUCT_SALE_STATUS.ONSALE,
         isSoldOut = false,
     }: GroupCodeParams): Promise<AxiosResponse> =>
         request({
@@ -197,7 +97,7 @@ const product = {
             totalReviewCount: false,
             familyMalls: false,
         },
-        order = { direction: DIRECTION.DESC },
+        order = { direction: ORDER_DIRECTION.DESC },
         categoryNos,
         brandNos,
         partnerNo,
