+++ src/models/promotion/index.ts
@@ -1,0 +1,9 @@
+export interface CouponsParams {
+    endYmd?: string;
+    pageNumber: number;
+    pageSize: number;
+    startYmd?: string;
+    usable: boolean | null;
+}
+
+export type TargetParams = Pick<CouponsParams, 'pageNumber' | 'pageSize'>;
